检索服务-Bug修改

+++ gulimall-search/src/main/java/com/atguigu/gulimall/search/vo/SearchResult.java
@@ -28,7 +28,7 @@ public class SearchResult {
     /**
      * 总页码
      */
-    private Integer totalPages;
+    private Long totalPages;
     /**
      * 当前查询的结果，所有涉及到的品牌
      */
